requirements.txt
beautifulsoup4==4.13.5
elasticsearch==9.1.1
fuzzywuzzy==0.18.0
langdetect==1.0.9
matplotlib==3.10.6
numpy==2.3.3
openai==1.107.2
pandas==2.3.2
pdf2image==1.17.0
Pillow==11.3.0
pyarrow==19.0.1
pypdf==6.0.0
PyPDF2==1.28.6
PyPDF2==3.0.1
python-dotenv==1.1.1
rapidfuzz==3.14.1
Requests==2.32.5
seaborn==0.13.2
selenium==4.35.0
tiktoken==0.9.0
tqdm==4.67.1
urllib3==2.5.0
